backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "pymongo>=4.10.1",
    "bcrypt>=4.2.1",
    "python-dotenv>=1.0.1",
    "PyJWT>=2.10.1",
    "huggingface_hub>=0.24.0",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "flask-cors>=4.0.0",
]
